requirements.txt
# Core dependencies for SILK implementation
# PyTorch - deep learning framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Numerical computing
numpy>=1.17.4

# Configuration management
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Training monitoring and logging
tensorboard>=2.12.0
tqdm>=4.65.0

# Utilities
scipy>=1.10.0

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.0.0

imageio>=2.0.0